server/pyproject.toml
[tool.poetry]
name = "lafontaine"
version = "0.1.0"
description = ""
authors = ["Madelaine Boyd <madelaine.boyd+github@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai2 = "^1.6.5"
python-docx = "^1.1.0"
python-magic = "^0.4.27"
boto3 = "^1.34.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
